requirements.txt
streamlit
faiss-cpu
sentence-transformers
langchain
duckduckgo-search
pdfplumber